+++ src/tabpfn/classifier.py
@@ -56,6 +56,7 @@ from tabpfn.preprocessing import (
     default_classifier_preprocessor_configs,
 )
 from tabpfn.utils import (
+    DevicesSpecification,
     fix_dtypes,
     get_embeddings,
     get_ordinal_encoder,
@@ -93,8 +94,13 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
     interface_config_: ModelInterfaceConfig
     """Additional configuration of the interface for expert users."""
 
-    device_: torch.device
-    """The device determined to be used."""
+    devices_: tuple[torch.device, ...]
+    """The devices determined to be used.
+
+    The devices are determined based on the `device` argument to the constructor, and
+    the devices available on the system. If multiple devices are listed, currently only
+    the first is used for inference.
+    """
 
     feature_names_in_: npt.NDArray[Any]
     """The feature names of the input data.
@@ -148,7 +154,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
         balance_probabilities: bool = False,
         average_before_softmax: bool = False,
         model_path: str | Path | Literal["auto"] = "auto",
-        device: str | torch.device | Literal["auto"] = "auto",
+        device: DevicesSpecification = "auto",
         ignore_pretraining_limits: bool = False,
         inference_precision: _dtype | Literal["autocast", "auto"] = "auto",
         fit_mode: Literal[
@@ -476,7 +482,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
         )
 
         check_cpu_warning(
-            self.device, X, allow_cpu_override=self.ignore_pretraining_limits
+            self.devices_, X, allow_cpu_override=self.ignore_pretraining_limits
         )
 
         if feature_names_in is not None:
@@ -600,7 +606,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
             byte_size, rng = self._initialize_model_variables()
         else:
             _, _, byte_size = determine_precision(
-                self.inference_precision, self.device_
+                self.inference_precision, self.devices_
             )
             rng = None
 
@@ -612,7 +618,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
             ensemble_configs=configs,
             cat_ix=cat_ix,
             fit_mode="batched",
-            device_=self.device_,
+            devices_=self.devices_,
             rng=rng,
             n_jobs=self.n_jobs,
             byte_size=byte_size,
@@ -647,7 +653,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
         else:  # already fitted and prompt_tuning mode: no cat. features
             _, rng = infer_random_state(self.random_state)
             _, _, byte_size = determine_precision(
-                self.inference_precision, self.device_
+                self.inference_precision, self.devices_
             )
 
         # Create the inference engine
@@ -658,7 +664,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
             ensemble_configs=ensemble_configs,
             cat_ix=self.inferred_categorical_indices_,
             fit_mode=self.fit_mode,
-            device_=self.device_,
+            devices_=self.devices_,
             rng=rng,
             n_jobs=self.n_jobs,
             byte_size=byte_size,
@@ -785,7 +791,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
     def _apply_balancing(self, probas: torch.Tensor) -> torch.Tensor:
         """Applies class balancing to a probability tensor."""
         class_prob_in_train = self.class_counts_ / self.class_counts_.sum()
-        balanced_probas = probas / torch.Tensor(class_prob_in_train).to(self.device_)
+        balanced_probas = probas / torch.Tensor(class_prob_in_train).to(probas.device)
         return balanced_probas / balanced_probas.sum(dim=-1, keepdim=True)
 
     def forward(  # noqa: C901, PLR0912
@@ -852,7 +858,7 @@ class TabPFNClassifier(ClassifierMixin, BaseEstimator):
         outputs = []
         for output, config in self.executor_.iter_outputs(
             X,
-            device=self.device_,
+            devices=self.devices_,
             autocast=self.use_autocast_,
         ):
             original_ndim = output.ndim