+++ tests/test_utils.py
@@ -3,11 +3,13 @@
 from __future__ import annotations
 
 import os
+from unittest.mock import MagicMock
 
 import numpy as np
 import pytest
+import torch
 
-from tabpfn.utils import infer_categorical_features
+from tabpfn.utils import infer_categorical_features, infer_devices
 
 
 def test_internal_windows_total_memory():
@@ -118,3 +120,77 @@ def test_infer_categorical_with_dict_raises_error():
             max_unique_for_category=2,
             min_unique_for_numerical=2,
         )
+
+
+def test__infer_devices__auto__cuda_and_mps_not_available__selects_cpu(
+    mocker: MagicMock,
+) -> None:
+    mocker.patch("torch.cuda").is_available.return_value = False
+    mocker.patch("torch.backends.mps").is_available.return_value = False
+    assert infer_devices(devices="auto") == (torch.device("cpu"),)
+
+
+def test__infer_devices__auto__single_cuda_gpu_available__selects_it(
+    mocker: MagicMock,
+) -> None:
+    mock_cuda = mocker.patch("torch.cuda")
+    mock_cuda.is_available.return_value = True
+    mock_cuda.device_count.return_value = 1
+    mocker.patch("torch.backends.mps").is_available.return_value = True
+    assert infer_devices(devices="auto") == (torch.device("cuda:0"),)
+
+
+def test__infer_devices__auto__multiple_cuda_gpus_available__selects_all(
+    mocker: MagicMock,
+) -> None:
+    mock_cuda = mocker.patch("torch.cuda")
+    mock_cuda.is_available.return_value = True
+    mock_cuda.device_count.return_value = 3
+    mocker.patch("torch.backends.mps").is_available.return_value = True
+
+    inferred = set(infer_devices(devices="auto"))
+    expected = {torch.device("cuda:0"), torch.device("cuda:1"), torch.device("cuda:2")}
+    assert inferred == expected
+
+
+def test__infer_devices__auto__cuda_and_mps_available_but_excluded__selects_cpu(
+    mocker: MagicMock, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    monkeypatch.setenv("TABPFN_EXCLUDE_DEVICES", "mps,cuda")
+    mock_cuda = mocker.patch("torch.cuda")
+    mock_cuda.is_available.return_value = True
+    mock_cuda.device_count.return_value = 1
+    mocker.patch("torch.backends.mps").is_available.return_value = True
+    assert infer_devices(devices="auto") == (torch.device("cpu"),)
+
+
+def test__infer_devices__device_specified__selects_it(
+    mocker: MagicMock,
+) -> None:
+    mock_cuda = mocker.patch("torch.cuda")
+    mock_cuda.is_available.return_value = True
+    mock_cuda.device_count.return_value = 2
+    mocker.patch("torch.backends.mps").is_available.return_value = True
+
+    assert infer_devices(devices="cuda:0") == (torch.device("cuda:0"),)
+
+
+def test__infer_devices__multiple_devices_specified___selects_them(
+    mocker: MagicMock,
+) -> None:
+    mock_cuda = mocker.patch("torch.cuda")
+    mock_cuda.is_available.return_value = True
+    mock_cuda.device_count.return_value = 3
+    mocker.patch("torch.backends.mps").is_available.return_value = False
+
+    inferred = set(infer_devices(devices=["cuda:0", "cuda:1", "cuda:4"]))
+    expected = {torch.device("cuda:0"), torch.device("cuda:1"), torch.device("cuda:4")}
+    assert inferred == expected
+
+
+def test__infer_devices__device_selected_twice__raises() -> None:
+    with pytest.raises(
+        ValueError,
+        match="The list of devices for inference cannot contain the same device more ",
+    ):
+        infer_devices(devices=["cpu", "cpu"])